Fixed MCTS and revamped environment.
I changed the environment to be 20x20 rather than 40x40 to simplify/speedup training. Moreover, I changed the input volume of the NN to incorporate the entire current map (instead of the visible parts of the map) and visible squares. The changes to the notebook reflect these changes to the environment and is used purely for testing the networks produced by parallelPPO.py.